Split physics
